Added function implementation.

+++ server/handler.go
@@ -6,16 +6,22 @@ import (
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/context"
-	"github.com/wst-libs/wst/logs"
 )
 
 func PutFileHandler(ctx *context.Context) []byte {
 	s, err := PutFileRequest(ctx.Input.RequestBody)
 	if err != nil {
-		return []byte("")
+		return JsonFormatErr()
 	}
+	info := FileInfo{
+		FilePath: s.Data.Path,
+		FileName: s.Data.Name,
+		FileType: s.Data.Type,
+		Bucket:   s.Data.Bucket,
+		Object:   s.Data.Object,
+	}
+	FileChan <- info
 
-	log.Printf("%v\n", s)
 	return PutFileResponse()
 }
 
@@ -27,24 +33,37 @@ func GetFileHandler(ctx *context.Context) []byte {
 	}
 	log.Println("raw query: ", u.RawQuery)
 	m, _ := url.ParseQuery(u.RawQuery)
-	log.Println(m)
-
-	aliyun, err := NewAliyunObject(beego.AppConfig.String("endpoint"), beego.AppConfig.String("accesskey"), beego.AppConfig.String("secretkey"), beego.AppConfig.String("bucket"))
-	if err != nil {
-		logs.Error("PostFileWithUrl error: ", err.Error())
-		return []byte{}
+	bucket := m.Get("bucket")
+	object := m.Get("object")
+	if len(bucket) <= 0 {
+		bucket = "llvision"
+	}
+	if len(object) <= 0 {
+		return UnknownReq()
 	}
 
-	isExist, err := aliyun.IsFileExist("filename")
+	obj, err := NewAliyunObject(beego.AppConfig.String("endpoint"), beego.AppConfig.String("accesskey"), beego.AppConfig.String("secretkey"), bucket)
 	if err != nil {
-		logs.Error("PostFileWithUrl Error: ", err.Error())
+		log.Println("GetFileHandler error: ", err.Error())
+		return BucketNotFound()
+	}
+
+	isExist, err := obj.IsFileExist(m.Get("object"))
+	if err != nil {
+		log.Println("GetFileHandler Error: ", err.Error())
+		return InternalError()
 	}
 	if isExist != true {
-
+		log.Println("file not exist.")
+		return FileNotFound()
 	}
-	returl, err := aliyun.GetFileWithURL("filename", 3600)
+	returl, err := obj.GetFileWithURL(object, 3600)
+	if err != nil {
+		log.Println("GetFileHandler error: ", err.Error())
+		return InternalError()
+	}
 	log.Println("url: ", returl)
-	return GetFileResponse()
+	return GetFileResponse(returl)
 }
 
 func UpdateFileHandler(ctx *context.Context) []byte {
@@ -56,6 +75,31 @@ func UpdateFileHandler(ctx *context.Context) []byte {
 	log.Println("raw query: ", u.RawQuery)
 	m, _ := url.ParseQuery(u.RawQuery)
 	log.Println(m)
+	bucket := m.Get("bucket")
+	object := m.Get("object")
+
+	s, err := UpdateFileRequest(ctx.Input.RequestBody)
+	if err != nil {
+		return JsonFormatErr()
+	}
+
+	obj, err := NewAliyunObject(beego.AppConfig.String("endpoint"), beego.AppConfig.String("accesskey"), beego.AppConfig.String("secretkey"), bucket)
+	if err != nil {
+		log.Println("UpdateFileHandler error: ", err.Error())
+		return BucketNotFound()
+	}
+
+	isExist, err := obj.IsFileExist(object)
+	if err != nil {
+		log.Println("UpdateFileHandler error: ", err.Error())
+		return InternalError()
+	}
+	if isExist != true {
+		log.Println("file not exist.")
+		return FileNotFound()
+	}
+
+	obj.UpdateFile(object, "description", s.Data.Desc)
 
 	return UpdateFileResponse()
 }
@@ -69,6 +113,26 @@ func DeleteFileHandler(ctx *context.Context) []byte {
 	log.Println("raw query: ", u.RawQuery)
 	m, _ := url.ParseQuery(u.RawQuery)
 	log.Println(m)
+	bucket := m.Get("bucket")
+	object := m.Get("object")
+
+	obj, err := NewAliyunObject(beego.AppConfig.String("endpoint"), beego.AppConfig.String("accesskey"), beego.AppConfig.String("secretkey"), bucket)
+	if err != nil {
+		log.Println("UpdateFileHandler error: ", err.Error())
+		return BucketNotFound()
+	}
+
+	isExist, err := obj.IsFileExist(object)
+	if err != nil {
+		log.Println("UpdateFileHandler error: ", err.Error())
+		return InternalError()
+	}
+	if isExist != true {
+		log.Println("file not exist.")
+		return FileNotFound()
+	}
+
+	obj.DeleteFile(object)
 
 	return DeleteFileResponse()
 }
